Refactor utils and tools into tools and display
