+++ paperoni/display.py
@@ -1,4 +1,3 @@
-import re
 import shutil
 import textwrap
 from typing import Union
@@ -36,25 +35,6 @@ link_generators = {
         "abstract": "https://www.semanticscholar.org/paper/{}"
     },
 }
-
-
-class QueryError(Exception):
-    pass
-
-
-class MutuallyExclusiveError(RuntimeError):
-    """Exception raised when mutually exclusive parameters are used in queries."""
-
-    def __init__(self, *args):
-        self.args = args
-
-    def __str__(self):
-        return "Mutually exclusive parameters: " + " vs ".join(
-            self._param_to_str(arg) for arg in self.args
-        )
-
-    def _param_to_str(self, param):
-        return param if isinstance(param, str) else f"({', '.join(param)})"
 
 
 def print_field(title, contents, bold=False):
@@ -177,33 +157,3 @@ def display(venue: Venue):
     print_field("Links", "")
     for typ, link in expand_links(venue.links):
         print(f"  {T.bold_green(typ):20} {link}")
-
-
-url_extractors = {
-    r"https?://[a-z.]*arxiv\.org/(?:abs|pdf)/([0-9]{4}\.[0-9]+).*": "arxiv",
-    r"https?://[a-z.]*arxiv-vanity\.com/papers/([0-9]{4}\.[0-9]+).*": "arxiv",
-    r"https?://(?:[^/]*)arxiv(?:[^/]*)\.cornell\.edu/abs/([0-9]{4}\.[0-9]+).*": "arxiv",
-    r"https?://scirate\.com/arxiv/([0-9]{4}\.[0-9]+).*": "arxiv",
-    r"https?://pubmed\.ncbi\.nlm\.nih\.gov/([^/]*)/": "pubmed",
-    r"https?://www\.ncbi\.nlm\.nih\.gov/pubmed/([^/]*)": "pubmed",
-    r"https?://www\.ncbi\.nlm\.nih\.gov/pmc/articles/([^/]*)": "pmc",
-    r"https?://europepmc.org/article/PMC/([^/]*)": "pmc",
-    r"https?://(?:dx\.)?doi\.org/(.*)": "doi",
-    r"https?://(?:www\.)?openreview\.net/(?:pdf\?|forum\?)id=(.*)": "openreview",
-    r"https?://dblp.uni-trier.de/db/([^/]+)/([^/]+)/[^/]+\.html#(.*)": "dblp",
-}
-
-
-def url_to_id(url):
-    for pattern, key in url_extractors.items():
-        if m := re.match(pattern, url):
-            lnk = "/".join(m.groups())
-            return (key, lnk)
-    return None
-
-
-def canonicalize_links(links):
-    links = {
-        url_to_id(url := link["link"]) or (link["type"], url) for link in links
-    }
-    return [{"type": typ, "link": lnk} for typ, lnk in links]
